requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
chromadb
youtube-transcript-api
python-dotenv
